requirements.txt
beautifulsoup4==4.13.4
bleach==6.2.0
html2text==2025.4.15
lxml==5.4.0
Markdown==3.8
networkx==3.4.2
numpy==2.2.6
obsidiantools==0.10.0
pandas==2.2.3
pymdown-extensions==10.15
python-dateutil==2.9.0.post0
python-frontmatter==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
webencodings==0.5.1
